DataSourceUtils.java の切り出し
